+++ src/app/components/side-nav.component.ts
@@ -10,6 +10,7 @@ export class SideNavState {
 }
 
 @Component({
+  selector: 'r-side-nav',
   style: css`
 		:host {
 			display: block;
@@ -248,7 +249,5 @@ class RSideNavComponent extends CustomElement {
   }
 
 }
-
-customElements.define('r-side-nav', RSideNavComponent);
 
 export { RSideNavComponent };