+++ src/app/lib/components/item/item.component.ts
@@ -29,6 +29,4 @@ class MyItemComponent extends CustomElement {
   }
 }
 
-customElements.define('my-item', MyItemComponent);
-
 export { MyItemComponent };