+++ src/app/lib/components/tree/node.component.ts
@@ -44,6 +44,4 @@ class NodeComponent extends PseudoElement {
   }
 }
 
-customElements.define('x-node', NodeComponent);
-
 export { NodeComponent };